// client/src/graphql-types/rx-gateway/globalTypes.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

//==============================================================
// START Enums and Input Objects
//==============================================================

export enum DigitalAccountStatus {
  ACTIVE = "ACTIVE",
  CANCELLED = "CANCELLED",
  CLOSED = "CLOSED",
  SETUP = "SETUP",
  UNKNOWN = "UNKNOWN",
}

/**
 * Represents a type of origin system this account belongs to.
 */
export enum DigitalAccountType {
  BROADBAND = "BROADBAND",
  KRAKEN = "KRAKEN",
  LPG = "LPG",
  OHM = "OHM",
  SAP = "SAP",
  UNKNOWN = "UNKNOWN",
  VPP = "VPP",
}

export enum DigitalServiceStatus {
  ACTIVE = "ACTIVE",
  CANCELLED = "CANCELLED",
  CLOSED = "CLOSED",
  SETUP = "SETUP",
  UNKNOWN = "UNKNOWN",
}

export enum DigitalServiceType {
  ELECTRICITY = "ELECTRICITY",
  GAS = "GAS",
  HOT_WATER = "HOT_WATER",
  LPG_EXCHANGE = "LPG_EXCHANGE",
  LPG_EXCHANGE_METERED = "LPG_EXCHANGE_METERED",
  LPG_INSITU = "LPG_INSITU",
  LPG_INSITU_METERED = "LPG_INSITU_METERED",
  LPG_METERED = "LPG_METERED",
  LPG_TERMINAL = "LPG_TERMINAL",
  NBN = "NBN",
  OHM = "OHM",
  OPTICOMM = "OPTICOMM",
  SOLAR = "SOLAR",
  UNKNOWN = "UNKNOWN",
  VOIP = "VOIP",
  VPP_BATTERY = "VPP_BATTERY",
}

export enum DigitalUserCustomerType {
  MULTI_SITE_PORTAL = "MULTI_SITE_PORTAL",
  RESIDENTIAL = "RESIDENTIAL",
  SMALL_MEDIUM_ENTERPRISE = "SMALL_MEDIUM_ENTERPRISE",
}

export enum FeatureStatus {
  ACTIVE = "ACTIVE",
  INACTIVE = "INACTIVE",
}

export enum FeatureType {
  BATTERY = "BATTERY",
  EV = "EV",
  INTERCOM = "INTERCOM",
  OHM = "OHM",
}

export enum ServiceType {
  ELECTRICITY = "ELECTRICITY",
  GAS = "GAS",
  HOT_WATER = "HOT_WATER",
}

//==============================================================
// END Enums and Input Objects
//==============================================================